gui: correct replacement of amp character in the wallet name for QMenu
The comment in the code regarding the use of an "&"
on a menu item is misleading. If a wallet name has an "&" in it,
it is not supposed to be interpreted as a hot-key, but it should be
shown as it is without replacing it to an underscore.

Github-Pull: gui#828
Rebased-From: 8233ee41ab9648cd0c3bd78bc2a8d692a54d9ea0